added book and user customer exceptions

+++ demo/src/main/java/com/qa/demo/services/BookService.java
@@ -4,7 +4,10 @@ package com.qa.demo.services;
 import java.util.List;
 import java.util.Optional;
 
+
 import org.springframework.stereotype.Service;
+
+import com.qa.demo.exceptions.BookNotFoundException;
 
 import com.qa.demo.persistence.domain.Book;
 import com.qa.demo.persistence.repos.LibraryRepo;
@@ -34,16 +37,24 @@ public class BookService {
 
     //Get an individual book by ISBN 
     public Book getBookByIsbn(Long isbn){
+
+        Book found = this.repo.findByIsbn(isbn);
+        if (found == null){
+            throw new BookNotFoundException("The book for "+isbn+ " is not found");
+        }
         return this.repo.findByIsbn(isbn);
+        
     }
 
     //Get a book by title
     public List<Book> findBookByTitle(String title){
+
         return this.repo.findByTitleContaining(title);
     }
 
     //Get a book by author
      public List<Book> findBookbyAuthor(String authors){
+   
         return this.repo.findByAuthorsContaining(authors);
      }
 
@@ -62,6 +73,10 @@ public class BookService {
     // removes a book by id
     public boolean removeBook(Long id){
        
+        if (!this.repo.existsById(id)){
+            throw new BookNotFoundException("Book "+id + " is not found");
+        }
+
         this.repo.deleteById(id);
         boolean exists = this.repo.existsById(id);
         return !exists;
